refactor: refactor the data pipeline
